Remove SSR configurations and prepare for static build

+++ src/app/components/dashboard/dashboard.component.ts
@@ -1,39 +1,42 @@
-import { Component, OnInit } from '@angular/core';
-
+import { Component, OnInit, Inject, PLATFORM_ID } from '@angular/core';
 import { TaskService } from '../../services/task.service';
 import { Router } from '@angular/router';
-
 import { debounceTime, Subject } from 'rxjs';
-
+import { isPlatformBrowser } from '@angular/common';
 
 @Component({
   selector: 'app-dashboard',
-  standalone:false,
+  standalone: false,
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.css'],
-
 })
 export class DashboardComponent implements OnInit {
   tasks: any[] = [];
-  userId:any;
-  searchSubject=new Subject();
-  sText:string='';
+  userId: any;
+  searchSubject = new Subject();
+  sText: string = '';
   backup: any[] = [];
 
-  constructor(private taskService: TaskService, private router: Router) {}
+  constructor(
+    private taskService: TaskService,
+    private router: Router,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
 
   ngOnInit() {
-    const userId = 1; // Replace with logged-in user ID
-    this.userId=localStorage.getItem('id');
+    if (isPlatformBrowser(this.platformId)) {
+      this.userId = localStorage.getItem('id');
+    }
+
     this.taskService.getAllTasks().subscribe((data: any) => {
       this.tasks = data;
-      this.backup=data;
-
+      this.backup = data;
     });
-    this.searchSubject.pipe(debounceTime(500)).subscribe((res:any)=>{
-      console.log("***res****",res);
-      this.tasks=this.backup.filter((item:any)=>item?.title.includes(res))
-    })
+
+    this.searchSubject.pipe(debounceTime(500)).subscribe((res: any) => {
+      console.log("***res****", res);
+      this.tasks = this.backup.filter((item: any) => item?.title.includes(res));
+    });
   }
 
   deleteTask(taskId: number) {
@@ -42,18 +45,23 @@ export class DashboardComponent implements OnInit {
       this.tasks = this.tasks.filter((task) => task.id !== taskId);
     });
   }
-  updateTask(taskId:any){
-    console.log("TAskId:::   ",taskId)
-    this.router.navigate(['add-task'], { queryParams: {id: taskId}});
+
+  updateTask(taskId: any) {
+    console.log("TaskId:::   ", taskId);
+    this.router.navigate(['add-task'], { queryParams: { id: taskId } });
   }
 
   logout() {
     alert('Logged out successfully');
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
     this.router.navigate(['/login']);
   }
+
   searchText(ev: any): void {
     console.log("*search**", ev.target.value);
-    this.searchSubject.next(this.sText)
+    this.sText = ev.target.value;
+    this.searchSubject.next(this.sText);
   }
 }
